feat: move dialogues to more logical place
